pyproject.toml
[project]
name = "gohphotos"
version = "0.1.0"
description = "Immich photo frame for Raspberry Pi with e-ink display"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.0",
    "python-dotenv>=1.2.1",
]

